fix: move all the pocket path into (main) and solve the header shown in _layout of (main)
