+++ app/(main)/_layout.jsx
@@ -2,7 +2,11 @@ import { Stack } from "expo-router";
 
 export default function MainLayout() {
   return (
-    <Stack>
+    <Stack
+      screenOptions={{
+        headerShown: false,
+      }}
+    >
       <Stack.Screen
         name="home"
         options={{
