+++ app/(main)/pocket/index.jsx
@@ -1,5 +1,5 @@
 import { Redirect } from "expo-router";
 
 export default function PocketIndex() {
-  return <Redirect href="/pocket/onboarding" />;
+  return <Redirect href="/(main)/pocket/onboarding" />;
 }
